tencentcloud/requirements.txt
tencentcloud-sdk-python
PySocks
tcloud-python-sdk
